refactor: convert to native esm
