+++ packages/cocoa-discord/src/base/Center.ts
@@ -2,13 +2,13 @@ import { Client, EmbedBuilder } from "discord.js";
 
 import chalk from "chalk";
 
-import { EmbedStyle } from "../main";
-import { MessageEvents } from "../message";
-import { SlashEvents } from "../slash";
-import { GlobalCommand } from "../slash/types";
+import { EmbedStyle } from "../main/index.js";
+import { MessageEvents } from "../message/index.js";
+import { SlashEvents } from "../slash/index.js";
+import { GlobalCommand } from "../slash/types.js";
 
-import { store } from "./store";
-import { Cog as BaseCog, BaseCommand, NonEmptyArray } from "./types";
+import { store } from "./store.js";
+import { Cog as BaseCog, BaseCommand, NonEmptyArray } from "./types.js";
 
 export abstract class ManagementCenter<
   Cog extends BaseCog<BaseCommand>,
