+++ packages/cocoa-discord/src/base/types.ts
@@ -1,4 +1,4 @@
-import { GlobalCommand } from "../slash";
+import { GlobalCommand } from "../slash/index.js";
 
 // * https://stackoverflow.com/a/49910890
 export type NonEmptyArray<T> = T[] & { 0: T };
