+++ packages/cocoa-discord/src/main/client.ts
@@ -2,7 +2,7 @@ import { Client } from "discord.js";
 
 import chalk from "chalk";
 
-import { store } from "../base";
+import { store } from "../base/index.js";
 
 /**
  * Do all the checks before login
