+++ packages/cocoa-discord/src/main/index.ts
@@ -1,8 +1,8 @@
-export * from "./activity";
-export * from "./client";
-export * from "./embed";
-export * from "./loader";
-export * from "./logger";
-export * from "./readline";
-export * from "./styles";
-export * from "./time";
+export * from "./activity.js";
+export * from "./client.js";
+export * from "./embed.js";
+export * from "./loader.js";
+export * from "./logger.js";
+export * from "./readline.js";
+export * from "./styles.js";
+export * from "./time.js";
