+++ packages/cocoa-discord/src/main/readline.ts
@@ -3,9 +3,9 @@ import { Client } from "discord.js";
 import chalk from "chalk";
 import readline from "readline";
 
-import { Awaitable } from "../base";
+import { Awaitable } from "../base/index.js";
 
-import { Loader } from "./loader";
+import { Loader } from "./loader.js";
 
 let rl: readline.Interface;
