+++ packages/cocoa-discord/src/slash/SlashSync.ts
@@ -8,9 +8,9 @@ import {
 
 import chalk from "chalk";
 
-import { getElapsed } from "../meta";
+import { getElapsed } from "../meta/index.js";
 
-import { GlobalCommand } from "./types";
+import { GlobalCommand } from "./types.js";
 
 /**
  * Represent the Command Data (discord.js-compatible interface)
