+++ packages/cocoa-discord/src/template/index.ts
@@ -1,3 +1,3 @@
-export * from "./builders";
-export * from "./options";
-export * from "./status";
+export * from "./builders.js";
+export * from "./options.js";
+export * from "./status.js";
